+++ R/compat-purrr.R
@@ -51,3 +51,20 @@ iwalk <- function(.x, .f, ...) {
   imap(.x, .f, ...)
   invisible(.x)
 }
+
+compact <- function(.x) {
+  Filter(length, .x)
+}
+
+transpose <- function (.l) {
+  inner_names <- names(.l[[1]])
+  if (is.null(inner_names)) {
+    fields <- seq_along(.l[[1]])
+  }
+  else {
+    fields <- rlang::set_names(inner_names)
+  }
+  map(fields, function(i) {
+    map(.l, .subset2, i)
+  })
+}